Reorganize for pypi publishing
* everything under bookserver directory will be in the package published to pypi
* Need a less generic name than app for the app directory
* config should be part of the package
