+++ pyproject.toml
@@ -1,15 +1,19 @@
 # ********************************
 # |docname| - Poetry configuration
 # ********************************
+#
+# See https://python-poetry.org/docs/dependency-specification/ to get an understanding of 
+# how poetry does specifies dependencies
+#
 [tool.poetry]
-name = "app"
+name = "bookserver"
 version = "0.1.0"
 description = "A new Runestone Server Framework"
 authors = ["Brad Miller <bonelake@mac.com>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.7"
 fastapi = "^0.62.0"
 uvicorn = "^0.13.1"
 Jinja2 = "^2.11.2"
